feat: Move functions into separate files
